producer/requirements.txt
pymongo==3.8.0
kafka-python==1.4.6
watchdog==0.9.0